reorganize components which are only used in one page
