+++ resources/js/app.ts
@@ -20,14 +20,30 @@ createInertiaApp({
         // e.g., 'settings.api-tokens' -> './pages/settings/ApiTokens.vue'
         // e.g., 'maps.edit' -> './pages/maps/Edit.vue'
         // e.g., 'maps.test' -> './pages/maps/Test.vue'
-        // e.g., 'ManageObjectTypes' -> './pages/manage-object-types/ManageObjectTypes.vue'
+        // e.g., 'auth/Login' -> './pages/auth/Login.vue'
+        // e.g., 'dashboard' -> './pages/dashboard/Dashboard.vue'
+        // e.g., 'welcome' -> './pages/welcome/Welcome.vue'
+
+        // Handle different naming patterns
+        if (name.includes('/')) {
+            // Direct path like 'auth/Login' -> './pages/auth/Login.vue'
+            return resolvePageComponent(`./pages/${name}.vue`, import.meta.glob<DefineComponent>('./pages/**/*.vue'));
+        }
 
         const parts = name.split('.');
         const directory = parts[0];
+
+        // Convert kebab-case to PascalCase for component name
         const componentName =
             parts.length > 1
-                ? parts[parts.length - 1].charAt(0).toUpperCase() + parts[parts.length - 1].slice(1)
-                : directory.charAt(0).toUpperCase() + directory.slice(1);
+                ? parts[parts.length - 1]
+                      .split('-')
+                      .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
+                      .join('')
+                : directory
+                      .split('-')
+                      .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
+                      .join('');
 
         const fullPath = `./pages/${directory}/${componentName}.vue`;
