+++ resources/js/layouts/app/AppHeaderLayout.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import AppContent from '@/components/AppContent.vue';
-import AppHeader from '@/components/AppHeader.vue';
 import AppShell from '@/components/AppShell.vue';
+import AppHeader from '@/layouts/app/partials/AppHeader.vue';
+import AppContent from '@/layouts/partials/AppContent.vue';
 import type { BreadcrumbItemType } from '@/types';
 
 interface Props {
