+++ resources/js/layouts/app/partials/AppSidebar.vue
@@ -3,10 +3,10 @@ import NavFooter from '@/components/NavFooter.vue';
 import NavMain from '@/components/NavMain.vue';
 import NavUser from '@/components/NavUser.vue';
 import { Sidebar, SidebarContent, SidebarFooter, SidebarHeader, SidebarMenu, SidebarMenuButton, SidebarMenuItem } from '@/components/ui/sidebar';
+import AppLogo from '@/layouts/partials/AppLogo.vue';
 import { type NavItem } from '@/types';
 import { Link } from '@inertiajs/vue3';
 import { Box, Folder, LayoutGrid, Palette } from 'lucide-vue-next';
-import AppLogo from './AppLogo.vue';
 
 const mainNavItems: NavItem[] = [
     {
