+++ resources/js/layouts/auth/AuthCardLayout.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import AppLogoIcon from '@/layouts/partials/AppLogoIcon.vue';
 import { Link } from '@inertiajs/vue3';
 
 defineProps<{
