+++ resources/js/layouts/partials/AppContent.vue
