+++ resources/js/layouts/partials/AppLogo.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import AppLogoIcon from '@/components/AppLogoIcon.vue';
+import AppLogoIcon from '@/layouts/partials/AppLogoIcon.vue';
 </script>
 
 <template>
