+++ resources/js/pages/manage-field-types/partials/FieldTypeList.vue
@@ -64,13 +64,13 @@
 </template>
 
 <script setup lang="ts">
-import CreateFieldTypeDialog from '@/components/CreateFieldTypeDialog.vue';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogContent, DialogTrigger } from '@/components/ui/dialog';
 import { FieldTypeService, type FieldType } from '@/services/FieldTypeService';
 import type { PageProps } from '@/types/globals';
 import { usePage } from '@inertiajs/vue3';
 import { onMounted, ref } from 'vue';
+import CreateFieldTypeDialog from './CreateFieldTypeDialog.vue';
 
 const page = usePage<PageProps>();
 const loading = ref(false);
