+++ resources/js/pages/maps/partials/BrushCursor.vue
