+++ resources/js/pages/maps/partials/EditorMiniMap.vue
