+++ resources/js/pages/maps/partials/EditorToolbar.vue
