+++ resources/js/pages/maps/partials/EraseCursor.vue
