+++ resources/js/pages/maps/partials/FieldTypeBox.vue
