+++ resources/js/pages/maps/partials/FieldTypeCursor.vue
