+++ resources/js/pages/maps/partials/FillCursor.vue
