+++ resources/js/pages/maps/partials/GridOverlay.vue
