+++ resources/js/pages/maps/partials/TileCanvas.vue
