+++ resources/js/pages/maps/partials/ToolCursor.vue
