+++ resources/js/pages/welcome/Welcome.vue
